backend/integration-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
alembic==1.12.1
redis==5.0.1
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
slowapi==0.1.8
cryptography==41.0.7
aiohttp==3.9.1
beautifulsoup4==4.12.2
pandas==2.1.3
numpy==1.26.2
playwright==1.40.0
python-telegram-bot==20.7
tweepy==4.14.0
facebook-sdk==3.1.0
google-auth==2.23.4
google-api-python-client==2.108.0
stripe==7.6.0
sendgrid==6.10.0
twilio==8.10.0
slack-sdk==3.26.0
discord.py==2.3.2
bcrypt==4.0.1
httpx==0.25.1 